Moved the spmm optimization code to a new header file
